# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'EzexStore API Documentation'
  description: 'API Documentation for EzexStore Backend System'
  version: 1.0.0
servers:
  -
    url: 'http://ezexstore-back.test'
paths:
...
  /api/inventory/adjust:
    post:
      summary: 'Adjust Stock'
      operationId: adjustStock
      description: "Adjusts the stock quantity of a product in a specific warehouse.\nUse type 'restock' to add inventory or 'order' to reduce it."
      parameters:
        -
          in: header
          name: X-API-KEY
          description: ''
          example: '{YOUR_API_KEY}'
          schema:
            type: string
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  data:
                    product_id: 1
                    warehouse_id: 1
                    quantity: 10
                    type: restock
                    created_at: '2024-12-18T00:00:00.000000Z'
                  message: 'Stock adjusted successfully'
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      product_id:
                        type: integer
                        example: 1
                      warehouse_id:
                        type: integer
                        example: 1
                      quantity:
                        type: integer
                        example: 10
                      type:
                        type: string
                        example: restock
                      created_at:
                        type: string
                        example: '2024-12-18T00:00:00.000000Z'
                  message:
                    type: string
                    example: 'Stock adjusted successfully'
        400:
          description: ''
          content:
            application/json:
              schema:
                oneOf:
                  -
                    description: 'insufficient stock'
                    type: object
                    example:
                      success: false
                      message: 'Insufficient stock. Available: 5, Required: 10'
                      error_code: insufficient_stock
                      details:
                        current_stock: 5
                        required_stock: 10
                        missing_stock: 5
                    properties:
                      success:
                        type: boolean
                        example: false
                      message:
                        type: string
                        example: 'Insufficient stock. Available: 5, Required: 10'
                      error_code:
                        type: string
                        example: insufficient_stock
                      details:
                        type: object
                        properties:
                          current_stock:
                            type: integer
                            example: 5
                          required_stock:
                            type: integer
                            example: 10
                          missing_stock:
                            type: integer
                            example: 5
                  -
                    description: 'invalid ownership'
                    type: object
                    example:
                      success: false
                      message: 'Invalid product or warehouse for this tenant'
                      error_code: invalid_ownership
                      details:
                        product_exists: false
                        warehouse_exists: true
                    properties:
                      success:
                        type: boolean
                        example: false
                      message:
                        type: string
                        example: 'Invalid product or warehouse for this tenant'
                      error_code:
                        type: string
                        example: invalid_ownership
                      details:
                        type: object
                        properties:
                          product_exists:
                            type: boolean
                            example: false
                          warehouse_exists:
                            type: boolean
                            example: true
      tags:
        - Inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: integer
                  description: 'The ID of the product.'
                  example: 1
                  nullable: false
                warehouse_id:
                  type: integer
                  description: 'The ID of the warehouse.'
                  example: 1
                  nullable: false
                quantity:
                  type: integer
                  description: 'The quantity to adjust (must be positive).'
                  example: 10
                  nullable: false
                type:
                  type: string
                  description: 'The type of adjustment (order, restock).'
                  example: restock
                  nullable: false
              required:
                - product_id
                - warehouse_id
                - quantity
                - type
  /api/inventory/stock:
    get:
      summary: 'Get Current Stock'
      operationId: getCurrentStock
      description: 'Returns the current stock quantity of a product in a specific warehouse.'
      parameters:
        -
          in: header
          name: X-API-KEY
          description: ''
          example: '{YOUR_API_KEY}'
          schema:
            type: string
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  data:
                    stock: 15
                  message: 'Stock retrieved successfully'
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      stock:
                        type: integer
                        example: 15
                  message:
                    type: string
                    example: 'Stock retrieved successfully'
        422:
          description: 'validation error'
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'The product id field is required.'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'The product id field is required.'
      tags:
        - Inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: integer
                  description: 'The ID of the product.'
                  example: 1
                  nullable: false
                warehouse_id:
                  type: integer
                  description: 'The ID of the warehouse.'
                  example: 1
                  nullable: false
              required:
                - product_id
                - warehouse_id
...
components:
  securitySchemes:
    default:
      type: apiKey
      name: X-API-KEY
      in: header
      description: 'Todas las peticiones deben incluir una API key válida en el header X-API-KEY.'
security:
  -
    default: []